Re-organize classes and constant in different files.
