+++ .gitignore
@@ -1,0 +1,2 @@
+
+game/__pycache__
