+++ game/constants.py
@@ -1,0 +1,20 @@
+import pygame
+
+pygame.init()
+
+BLANK_INT_BOARD = [[0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [0, 0, 0, 0, 0, 0, 0, 0, 0]]
+BACKGROUND_COLOR = (64, 64, 64)
+CELL_COLOR = "#c0e8ec"
+TEXT_COLOR = (64, 64, 64)
+NUM_FONT = pygame.font.Font(None, 60)
+TEXT_FONT = pygame.font.Font(None, 50)
+CLEAR = False
+SOLVE = False
